Recorded encoder data at practice
Had some connection issues with PDH that ate up time. I was able to get wrist encoder values for some key positions (see the .md) I made button reassignments after talking with Cole. Everything seems to be working except for wrist. We'd like to put the wrist actions on the right thumbstick. I need to work on the Stow commands and Load commands. Didn't get a chance to test Auton. Tuesday is programming practice. Wednesday is drive practice. Adding some of those in constants.java. Those numbers will need to be tweaked, but should be good enough for now.